% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MediaInfo.R
\name{MediaInfo}
\alias{MediaInfo}
\title{Retrieve media information}
\usage{
MediaInfo(filepath, MediaInfo_Path)
}
\arguments{
\item{filepath}{file path of media file}

\item{MediaInfo_Path}{Path to MediaInfo.exe}
}
\description{
Uses program MediaInfo to extract details from media (e.g. video or image files).
}
\details{
You can download MediaInfo from https://mediaarea.net/en/MediaInfo
}
