% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MediaInfo.R
\name{parse_video_fps}
\alias{parse_video_fps}
\title{Parse frames per second information}
\usage{
parse_video_fps(vid_info)
}
\arguments{
\item{vid_info}{video metadata output from the MediaInfo function}
}
\description{
Parses frames per second of video from output of MediaInfo function
}
