% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MediaInfo.R
\name{parse_video_time}
\alias{parse_video_time}
\title{Parse video duration}
\usage{
parse_video_time(vid_info)
}
\arguments{
\item{vid_info}{output from MediaInfo function}
}
\description{
Parses duration of video in seconds from output of MediaInfo function
}
